% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghl.R
\name{function_hl}
\alias{function_hl}
\title{wrapper function to extract method from existing Stat and assign data to be highlighted}
\usage{
function_hl(nm, stat)
}
\arguments{
\item{nm}{name of method in the ggproto object to extract}

\item{stat}{ggproto object of class Stat to inherit from}
}
\value{
function
}
\description{
wrapper function to extract method from existing Stat and assign data to be highlighted
}
