% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghl.R
\name{function_ll}
\alias{function_ll}
\title{wrapper function to extract method from existing Stat and assign data to be lowlighted}
\usage{
function_ll(nm, stat, LL)
}
\arguments{
\item{nm}{name of method in the ggproto object to extract}

\item{stat}{ggproto object of class Stat to inherit from}

\item{LL}{list of aethentics of lowlight objects (e.g., list(colour = NA))}
}
\value{
function
}
\description{
wrapper function to extract method from existing Stat and assign data to be lowlighted
}
