% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghl.R
\name{geom_line_hl}
\alias{geom_line_hl}
\title{highlighted version of ggplot2::geom_line}
\usage{
geom_line_hl(..., LL = .LL)
}
\arguments{
\item{...}{arguments passed to geom_point}

\item{LL}{aethentics for lowlights specified by list (e.g., list(colour = 'gray', alpha = '0.5'))}
}
\description{
highlighted version of ggplot2::geom_line
}
