% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghl.R
\name{gghl}
\alias{gghl}
\title{Construct highlighted version of geom_*.}
\usage{
gghl(geom, LL = list(colour = NA))
}
\arguments{
\item{geom}{geom_* functions (e.g., geom_point)}

\item{LL}{aethentics for lowlights specified by list (e.g., list(colour = 'gray', alpha = '0.5'))}
}
\value{
function
}
\description{
This function can highlight geom functions who have 'stat' as a parameter.
For example, geom_point, geom_rect, are acceptable.
However, geom_abline isn't.
In addition, geom_smooth does not workds currently.
}
\examples{
library(ggplot2)
geom_point_hl2 <- gghl(geom_point)
d <- data.frame(x = 1:5, y = 1, hl = 1:5 == 3)
ggplot(
  d,
  aes(x, y, highlight = hl)
) +
  geom_point_hl2()
ggplot(
  d,
  aes(x, y, highlight = hl, color = hl)
) +
  gghl(geom_point)()
}
