% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gginteractive.R
\name{ggplot_add.gginteractive}
\alias{ggplot_add.gginteractive}
\title{ggplot_add for gginteractive class}
\usage{
ggplot_add.gginteractive(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add \code{object} to}

\item{object_name}{The name of the object to add}
}
\value{
A modified ggplot object
}
\description{
This generic allows you to add your own methods for adding custom objects to
a ggplot with \link{+.gg}.
}
