% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghl.R
\name{ggproto_hl}
\alias{ggproto_hl}
\title{Return highlighted version of ggproto object a stat of geom_.}
\usage{
ggproto_hl(geom)
}
\arguments{
\item{geom}{geom_* functions (e.g., geom_point)}
}
\value{
ggproto object
}
\description{
Return highlighted version of ggproto object a stat of geom_.
}
