% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghl.R
\name{ggproto_ll}
\alias{ggproto_ll}
\title{Return lowlighted version of ggproto object a stat of geom_.}
\usage{
ggproto_ll(geom, LL)
}
\arguments{
\item{geom}{geom_* functions (e.g., geom_point)}

\item{LL}{aethentics for lowlights specified by list (e.g., list(colour = 'gray', alpha = '0.5'))}
}
\value{
ggproto object
}
\description{
Return lowlighted version of ggproto object a stat of geom_.
}
