% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeller.R
\name{label_tex}
\alias{label_tex}
\title{Labeller functions to parse TeX syntax}
\usage{
label_tex(labels, ...)
}
\arguments{
\item{labels}{Data frame of labels. Usually contains only one
element, but faceting over multiple factors entails multiple
label variables.}

\item{...}{Arguments passed on to \code{ggplot2::label_parsed}
\describe{
  \item{multi_line}{Whether to display the labels of multiple factors
on separate lines.}
}}
}
\description{
Labeller functions to parse TeX syntax
}
