% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_gg}
\alias{map_gg}
\alias{map_gg.data.frame}
\alias{map_gg.list}
\title{Apply \code{ggplot()} on a list of data frames or on a (grouped) data frame.}
\usage{
map_gg(data = NULL, ...)

\method{map_gg}{data.frame}(data = NULL, ...)

\method{map_gg}{list}(data = NULL, ...)
}
\arguments{
\item{data}{A list of data frames or a (grouped) data frame.}

\item{...}{Layers to be added to ggplot}
}
\value{
A \code{gglist} class object which is a list of \code{gg} class objects
}
\description{
Apply \code{ggplot()} on a list of data frames or on a (grouped) data frame.
}
