% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_gglist.R
\name{print_gglist}
\alias{print_gglist}
\title{print list of ggplot objects}
\usage{
print_gglist(gglist, ...)
}
\arguments{
\item{gglist}{list of ggplot objects}

\item{...}{other arguments passed to pforeach}
}
\description{
print list of ggplot objects
}
