% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ci.R
\name{stat_ci}
\alias{stat_ci}
\title{Confidence interval of smoothed density estimates}
\usage{
stat_ci(mapping = NULL, ..., ci = 0.95)
}
\arguments{
\item{mapping}{NULL}

\item{...}{other arguments described in stat_density}

\item{ci}{confidence interval. 0.95 in default}
}
\description{
Confidence interval of smoothed density estimates
}
\examples{
library(ggplot2)
ggplot(data.frame(x = rnorm(100)), aes(x)) +
stat_density() +
stat_ci(fill = 'orange')
}
