% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_error.R
\name{stat_err}
\alias{stat_err}
\title{stat_error
A function to draw both vertical and horizontal}
\usage{
stat_err(mapping = NULL, data = NULL, geom = "segment",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{NULL}

\item{data}{NULL}

\item{geom}{'segment'}

\item{position}{'identitiy'}

\item{na.rm}{FALSE}

\item{show.legend}{NA}

\item{inherit.aes}{TRUE}

\item{...}{other arguments passed to ggplot2::layer}
}
\description{
stat_error
A function to draw both vertical and horizontal
}
