% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_error.R
\docType{data}
\name{StatErr}
\alias{StatErr}
\title{StatErr}
\format{An object of class \code{StatErr} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatErr
}
\description{
StatErr
}
\keyword{datasets}
