% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghl.R
\name{geom_line_hl}
\alias{geom_line_hl}
\title{highlighted version of ggplot2::geom_line}
\usage{
geom_line_hl(...)
}
\arguments{
\item{...}{arguments passed to geom_point}
}
\description{
highlighted version of ggplot2::geom_line
}
