% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghl.R
\name{geom_point_hl}
\alias{geom_point_hl}
\title{highlighted version of ggplot2::geom_point}
\usage{
geom_point_hl(..., LL = .LL)
}
\arguments{
\item{...}{arguments passed to geom_point}
}
\description{
highlighted version of ggplot2::geom_point
}
