% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghl.R
\name{ggproto_ll}
\alias{ggproto_ll}
\title{Return lowlighted version of ggproto object a stat of geom_.}
\usage{
ggproto_ll(geom, LL)
}
\arguments{
\item{geom}{geom_* functions (e.g., geom_point)}

\item{LL}{list of aethentics of lowlight objects (e.g., list(colour = NA))}
}
\value{
ggproto object
}
\description{
Return lowlighted version of ggproto object a stat of geom_.
}
