% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_gglist.R
\name{print_gglist}
\alias{print_gglist}
\title{print list of ggplot objects}
\usage{
print_gglist(x, strategy = multiprocess, ...)
}
\arguments{
\item{x}{List of ggplot objects}

\item{strategy}{See \code{strategy} in `...``}

\item{...}{Arguments passed on to \code{future::plan}
\describe{
  \item{strategy}{The evaluation function (or name of it) to use
for resolving a future.  If NULL, then the current strategy is returned.}
  \item{substitute}{If TRUE, the \code{strategy} expression is
\code{substitute()}:d, otherwise not.}
  \item{.skip}{(internal) If \code{TRUE}, then attempts to set a strategy
that is the same as what is currently in use, will skipped.}
  \item{.call}{(internal) Used for recording the call to this function.}
  \item{.cleanup}{(internal) Used to stop implicitly started clusters.}
  \item{.init}{(internal) Used to initiate workers.}
}}
}
\description{
print list of ggplot objects
}
